backend/seo-analyze/requirements.txt
pydantic==2.5.0
openai==0.28.0
psycopg2-binary==2.9.9